fix(netcode): cast MsgMissionRoster gen to uint16_t on assign
MSVC treats the implicit uint32_t (EntityId::generation) -> uint16_t
(MsgMissionRoster::entityGen) narrowing as C4244, which /WX turns into an error
on the Windows CI leg (GCC/Clang accept it silently, so it passed locally). Add
the explicit static_cast at both roster send sites, matching the established
CombatEventRecord subjectGen/instigatorGen pattern. Part of #914.

+++ engine/net/WorldBroadcaster.cpp
@@ -313,7 +313,7 @@ void WorldBroadcaster::updateMissionRoster(const std::string& missionObjectId, E
         return; // unknown object, or the slot was freed (invalid entity) — nothing to advertise
     MsgMissionRoster rmsg{};
     rmsg.entityIdx = entity.index;
-    rmsg.entityGen = entity.generation;
+    rmsg.entityGen = static_cast<uint16_t>(entity.generation);
     std::snprintf(rmsg.objectId, sizeof(rmsg.objectId), "%s", missionObjectId.c_str());
     for (const auto& [peerId, pin] : m_peerInputs) {
         (void)pin;
@@ -5065,7 +5065,7 @@ void WorldBroadcaster::sendConnectAck(uint32_t peerId, EntityId assigned, PeerRo
                 continue; // invalid entity (e.g. an unbound player slot)
             MsgMissionRoster rmsg{};
             rmsg.entityIdx = eid.index;
-            rmsg.entityGen = eid.generation;
+            rmsg.entityGen = static_cast<uint16_t>(eid.generation);
             std::snprintf(rmsg.objectId, sizeof(rmsg.objectId), "%s", objectId.c_str());
             appendMsg(rbuf, rmsg);
         }
